# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /safety:
    get:
      operationId: listAlerts
      tags: [safety]
      summary: List community safety alerts
      parameters:
        - name: communityId
          in: query
          required: false
          schema:
            type: string
            default: "default"
        - name: resolved
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
    post:
      operationId: createAlert
      tags: [safety]
      summary: Report a safety alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertInput"
      responses:
        "201":
          description: Created alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /safety/{alertId}/resolve:
    patch:
      operationId: resolveAlert
      tags: [safety]
      summary: Mark alert as resolved
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /safety/{alertId}/comments:
    get:
      operationId: listAlertComments
      tags: [safety]
      summary: Get comments for an alert
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlertComment"
    post:
      operationId: createAlertComment
      tags: [safety]
      summary: Add a comment to an alert
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertCommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertComment"
  /marketplace:
    get:
      operationId: listListings
      tags: [marketplace]
      summary: List marketplace listings
      parameters:
        - name: communityId
          in: query
          required: false
          schema:
            type: string
            default: "default"
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: minPrice
          in: query
          required: false
          schema:
            type: integer
        - name: maxPrice
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 24
      responses:
        "200":
          description: List of listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingList"
    post:
      operationId: createListing
      tags: [marketplace]
      summary: Create a marketplace listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingInput"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /marketplace/{listingId}:
    get:
      operationId: getListing
      tags: [marketplace]
      summary: Get a single listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          description: Not found
  /marketplace/{listingId}/status:
    patch:
      operationId: updateListingStatus
      tags: [marketplace]
      summary: Update listing status (sold / reserved / available)
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingStatusUpdate"
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      parameters:
        - name: communityId
          in: query
          required: false
          schema: { type: string, default: "default" }
      responses:
        "200":
          description: Upcoming and past events
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventList" }
    post:
      operationId: createEvent
      tags: [events]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventInput" }
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
  /events/{eventId}/rsvp:
    post:
      operationId: rsvpEvent
      tags: [events]
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RsvpInput" }
      responses:
        "200":
          description: RSVP recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Rsvp" }
  /events/{eventId}/rsvps:
    get:
      operationId: listEventRsvps
      tags: [events]
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: RSVPs for event
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Rsvp" }
...
components:
  schemas:
...
    Alert:
      type: object
      required:
        - id
        - communityId
        - userId
        - userName
        - unitNumber
        - type
        - title
        - description
        - locationDetail
        - severity
        - isResolved
        - createdAt
      properties:
        id:
          type: integer
        communityId:
          type: string
        userId:
          type: string
        userName:
          type: string
        unitNumber:
          type: string
        type:
          type: string
        title:
          type: string
        description:
          type: string
        locationDetail:
          type: string
        imageUrl:
          type: string
          nullable: true
        severity:
          type: string
        isResolved:
          type: boolean
        createdAt:
          type: string
    AlertInput:
      type: object
      required:
        - type
        - title
        - description
      properties:
        communityId:
          type: string
          default: "default"
        userId:
          type: string
          default: "ahmed"
        userName:
          type: string
          default: "Ahmed Khan"
        unitNumber:
          type: string
          default: "B-204"
        type:
          type: string
        title:
          type: string
        description:
          type: string
        locationDetail:
          type: string
          default: ""
        imageUrl:
          type: string
          nullable: true
        severity:
          type: string
          default: "medium"
    AlertComment:
      type: object
      required:
        - id
        - alertId
        - userId
        - userName
        - unitNumber
        - body
        - createdAt
      properties:
        id:
          type: integer
        alertId:
          type: integer
        userId:
          type: string
        userName:
          type: string
        unitNumber:
          type: string
        body:
          type: string
        createdAt:
          type: string
    AlertCommentInput:
      type: object
      required:
        - body
      properties:
        userId:
          type: string
          default: "ahmed"
        userName:
          type: string
          default: "Ahmed Khan"
        unitNumber:
          type: string
          default: "B-204"
        body:
          type: string
    Listing:
      type: object
      required:
        - id
        - communityId
        - userId
        - userName
        - unitNumber
        - title
        - description
        - category
        - imageUrls
        - condition
        - status
        - whatsappNumber
        - createdAt
      properties:
        id:
          type: integer
        communityId:
          type: string
        userId:
          type: string
        userName:
          type: string
        unitNumber:
          type: string
        title:
          type: string
        description:
          type: string
        pricePkr:
          type: integer
          nullable: true
        category:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        condition:
          type: string
        status:
          type: string
        whatsappNumber:
          type: string
        createdAt:
          type: string
    ListingList:
      type: object
      required:
        - listings
        - total
        - page
        - limit
        - hasMore
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        hasMore:
          type: boolean
    ListingInput:
      type: object
      required:
        - title
        - description
        - whatsappNumber
      properties:
        communityId:
          type: string
          default: "default"
        userId:
          type: string
          default: "ahmed"
        userName:
          type: string
          default: "Ahmed Khan"
        unitNumber:
          type: string
          default: "B-204"
        title:
          type: string
        description:
          type: string
        pricePkr:
          type: integer
          nullable: true
        category:
          type: string
          default: "other"
        imageUrls:
          type: array
          items:
            type: string
        condition:
          type: string
          default: "good"
        whatsappNumber:
          type: string
    ListingStatusUpdate:
      type: object
      required:
        - status
      properties:
        status:
          type: string
    Event:
      type: object
      required: [id, communityId, userId, userName, unitNumber, title, description, date, time, location, rsvpCount, createdAt]
      properties:
        id: { type: integer }
        communityId: { type: string }
        userId: { type: string }
        userName: { type: string }
        unitNumber: { type: string }
        title: { type: string }
        description: { type: string }
        date: { type: string }
        time: { type: string }
        location: { type: string }
        imageUrl: { type: string, nullable: true }
        rsvpCount: { type: integer }
        createdAt: { type: string }
    EventInput:
      type: object
      required: [title, date]
      properties:
        communityId: { type: string, default: "default" }
        userId: { type: string, default: "ahmed" }
        userName: { type: string, default: "Ahmed Khan" }
        unitNumber: { type: string, default: "B-204" }
        title: { type: string }
        description: { type: string, default: "" }
        date: { type: string }
        time: { type: string, default: "" }
        location: { type: string, default: "" }
        imageUrl: { type: string, nullable: true }
    EventList:
      type: object
      required: [upcoming, past]
      properties:
        upcoming:
          type: array
          items: { $ref: "#/components/schemas/Event" }
        past:
          type: array
          items: { $ref: "#/components/schemas/Event" }
    Rsvp:
      type: object
      required: [id, eventId, userId, userName, status, createdAt]
      properties:
        id: { type: integer }
        eventId: { type: integer }
        userId: { type: string }
        userName: { type: string }
        status: { type: string }
        createdAt: { type: string }
    RsvpInput:
      type: object
      required: [status]
      properties:
        userId: { type: string, default: "ahmed" }
        userName: { type: string, default: "Ahmed Khan" }
        status: { type: string }
...