# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Alert {
  id: number;
  communityId: string;
  userId: string;
  userName: string;
  unitNumber: string;
  type: string;
  title: string;
  description: string;
  locationDetail: string;
  imageUrl?: string | null;
  severity: string;
  isResolved: boolean;
  createdAt: string;
}