# lib/api-zod/src/generated/types/rsvp.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Rsvp {
  id: number;
  eventId: number;
  userId: string;
  userName: string;
  status: string;
  createdAt: string;
}